Makes the AI now check if they can win
The AI now has the ability to check if they can win in the next turn and will do this rather than blocking the player from making a move. However there is still some flaws in the AI as it does not check for all future moves. A way to make this AI player unbeatable will be to implement a Minimax algorithm.